mpegvideoparse2: don't unnecessarily finish frame after seek

+++ gst/mpegvideoparse2/gstmpegvideoparse2.h
@@ -47,7 +47,7 @@ struct _GstMpegVideoParse2 {
   GstBaseVideoParse parent;
 
   GstMVP2State state;
-  guint8 prev_packet;
+  gint prev_packet;
   guint64 gop_start;
 
   GstBuffer *seq_header_buffer;
